[photons_app] moving task specific code around
