+++ modules/photons_app/photons_app.py
@@ -6,7 +6,7 @@ The specifications are responsible for sanitation, validation and normalisation.
 from photons_app.registers import TargetRegister, Target, ReferenceResolerRegister
 from photons_app.formatter import MergedOptionStringFormatter
 from photons_app.errors import BadOption, ApplicationStopped
-from photons_app.task_specifier import task_specifier_spec
+from photons_app.tasks import task_specifier_spec
 from photons_app import helpers as hp
 
 from delfick_project.norms import sb, dictobj, va
